feat:
Refactored App to be a functional component.

Added react-select and created the dropdown with names for login.

Added more tests; for AuthedUsers, Questions, Users.

Created a menu (with links to the different views/pages)

Changed RequireAuth component to redirect to the page the user wanted to access before being redirect to login

Added Logout page/functionality

Added basic initial css